Prepare Laravel OAS contract testing
This change adds a laravel plugin that tests API requests and responses
for their compatibility with the OpenAPI spec, aka "contract testing".

An additional "Contract" directory holds the contract test files and a
first example for the '/api/mismatches' endpoint has been added. Changes
needed in the OAS spec are applied to  the new mismatchfinder-api-spec.yaml
file.

Bug: T296825